Enables Prisma Data Proxy for all environments
Enables Prisma Data Proxy in development, build, and postinstall scripts.

Renames the test setup file and updates the vitest configuration to
reflect this change.

Configures vitest to ignore source map errors and sets the Node environment
variable for test environments.

Disables sourcemaps for tests and updates test coverage configuration
to be more verbose and not silent.